acpi: move parser opcode table into the parser pkg
