--- src/gopheros/device/acpi/aml/opcode.go
@@ -1,269 +1,0 @@
-package aml
-
-// opcode describes an AML opcode. While AML supports 256 opcodes, some of them
-// are specified using a combination of an extension prefix and a code. To map
-// each opcode into a single unique value the parser uses an uint16
-// representation of the opcode values.
-type opcode uint16
-
-// String implements fmt.Stringer for opcode.
-func (op opcode) String() string {
-	for _, entry := range opcodeTable {
-		if entry.op == op {
-			return entry.name
-		}
-	}
-
-	return "unknown"
-}
-
-// opIsLocalArg returns true if this opcode represents any of the supported local
-// function args 0 to 7.
-func opIsLocalArg(op opcode) bool {
-	return op >= opLocal0 && op <= opLocal7
-}
-
-// opIsMethodArg returns true if this opcode represents any of the supported
-// input function args 0 to 6.
-func opIsMethodArg(op opcode) bool {
-	return op >= opArg0 && op <= opArg6
-}
-
-// opIsArg returns true if this opcode is either a local or a method arg.
-func opIsArg(op opcode) bool {
-	return opIsLocalArg(op) || opIsMethodArg(op)
-}
-
-// opIsDataObject returns true if this opcode is part of a DataObject definition
-//
-// Grammar:
-// DataObject := ComputationalData | DefPackage | DefVarPackage
-// ComputationalData := ByteConst | WordConst | DWordConst | QWordConst | String | ConstObj | RevisionOp | DefBuffer
-// ConstObj := ZeroOp | OneOp | OnesOp
-func opIsDataObject(op opcode) bool {
-	switch op {
-	case opBytePrefix, opWordPrefix, opDwordPrefix, opQwordPrefix, opStringPrefix,
-		opZero, opOne, opOnes, opRevision, opBuffer, opPackage, opVarPackage:
-		return true
-	default:
-		return false
-	}
-}
-
-// opIsType2 returns true if this is a Type2Opcode.
-//
-// Grammar:
-// Type2Opcode := DefAcquire | DefAdd | DefAnd | DefBuffer | DefConcat |
-//  DefConcatRes | DefCondRefOf | DefCopyObject | DefDecrement |
-//  DefDerefOf | DefDivide | DefFindSetLeftBit | DefFindSetRightBit |
-//  DefFromBCD | DefIncrement | DefIndex | DefLAnd | DefLEqual |
-//  DefLGreater | DefLGreaterEqual | DefLLess | DefLLessEqual | DefMid |
-//  DefLNot | DefLNotEqual | DefLoadTable | DefLOr | DefMatch | DefMod |
-//  DefMultiply | DefNAnd | DefNOr | DefNot | DefObjectType | DefOr |
-//  DefPackage | DefVarPackage | DefRefOf | DefShiftLeft | DefShiftRight |
-//  DefSizeOf | DefStore | DefSubtract | DefTimer | DefToBCD | DefToBuffer |
-//  DefToDecimalString | DefToHexString | DefToInteger | DefToString |
-//  DefWait | DefXOr
-func opIsType2(op opcode) bool {
-	switch op {
-	case opAcquire, opAdd, opAnd, opBuffer, opConcat,
-		opConcatRes, opCondRefOf, opCopyObject, opDecrement,
-		opDerefOf, opDivide, opFindSetLeftBit, opFindSetRightBit,
-		opFromBCD, opIncrement, opIndex, opLand, opLEqual,
-		opLGreater, opLLess, opMid,
-		opLnot, opLoadTable, opLor, opMatch, opMod,
-		opMultiply, opNand, opNor, opNot, opObjectType, opOr,
-		opPackage, opVarPackage, opRefOf, opShiftLeft, opShiftRight,
-		opSizeOf, opStore, opSubtract, opTimer, opToBCD, opToBuffer,
-		opToDecimalString, opToHexString, opToInteger, opToString,
-		opWait, opXor:
-		return true
-	default:
-		return false
-	}
-}
-
-// opIsBufferField returens true if this opcode describes a
-// buffer field creation operation.
-func opIsBufferField(op opcode) bool {
-	switch op {
-	case opCreateField, opCreateBitField, opCreateByteField, opCreateWordField, opCreateDWordField, opCreateQWordField:
-		return true
-	default:
-		return false
-	}
-}
-
-// objType represents the object types that are supported by the AML parser.
-type objType uint8
-
-// The list of AML object types.
-const (
-	objTypeAny objType = iota
-	objTypeInteger
-	objTypeString
-	objTypeBuffer
-	objTypePackage
-	objTypeDevice
-	objTypeEvent
-	objTypeMethod
-	objTypeMutex
-	objTypeRegion
-	objTypePower
-	objTypeProcessor
-	objTypeThermal
-	objTypeBufferField
-	objTypeLocalRegionField
-	objTypeLocalBankField
-	objTypeLocalReference
-	objTypeLocalAlias
-	objTypeLocalScope
-	objTypeLocalVariable
-	objTypeMethodArgument
-)
-
-// opFlag specifies a list of OR-able flags that describe the object
-// type/attributes generated by a particular opcode.
-type opFlag uint16
-
-const (
-	opFlagNone opFlag = 1 << iota
-	opFlagHasPkgLen
-	opFlagNamed
-	opFlagConstant
-	opFlagReference
-	opFlagArithmetic
-	opFlagCreate
-	opFlagReturn
-	opFlagExecutable
-	opFlagNoOp
-	opFlagScoped
-)
-
-// is returns true if f is set in this opFlag.
-func (fl opFlag) is(f opFlag) bool {
-	return (fl & f) != 0
-}
-
-// opArgFlags encodes up to 7 opArgFlag values in a uint64 value.
-type opArgFlags uint64
-
-// argCount returns the number of encoded args in the given flag.
-func (fl opArgFlags) argCount() (count uint8) {
-	// Each argument is specified using 8 bits with 0x0 indicating the end of the
-	// argument list
-	for ; fl&0xf != 0; fl, count = fl>>8, count+1 {
-	}
-
-	return count
-}
-
-// arg returns the arg flags for argument "num" where num is the 0-based index
-// of the argument to return. The allowed values for num are 0-6.
-func (fl opArgFlags) arg(num uint8) opArgFlag {
-	return opArgFlag((fl >> (num * 8)) & 0xf)
-}
-
-// contains returns true if the arg flags contain any argument with type x.
-func (fl opArgFlags) contains(x opArgFlag) bool {
-	// Each argument is specified using 8 bits with 0x0 indicating the end of the
-	// argument list
-	for ; fl&0xf != 0; fl >>= 8 {
-		if opArgFlag(fl&0xf) == x {
-			return true
-		}
-	}
-
-	return false
-}
-
-// opArgFlag represents the type of an argument expected by a particular opcode.
-type opArgFlag uint8
-
-// The list of supported opArgFlag values.
-const (
-	_ opArgFlag = iota
-	opArgTermList
-	opArgTermObj
-	opArgByteList
-	opArgPackage
-	opArgString
-	opArgByteData
-	opArgWord
-	opArgDword
-	opArgQword
-	opArgNameString
-	opArgSuperName
-	opArgSimpleName
-	opArgDataRefObj
-	opArgTarget
-	opArgFieldList
-)
-
-// String implements fmt.Stringer for opArgFlag.
-func (fl opArgFlag) String() string {
-	switch fl {
-	case opArgTermList:
-		return "opArgTermList"
-	case opArgTermObj:
-		return "opArgTermObj"
-	case opArgByteList:
-		return "opArgByteList"
-	case opArgPackage:
-		return "opArgPackage"
-	case opArgString:
-		return "opArgString"
-	case opArgByteData:
-		return "opArgByteData"
-	case opArgWord:
-		return "opArgWord"
-	case opArgDword:
-		return "opArgDword"
-	case opArgQword:
-		return "opArgQword"
-	case opArgNameString:
-		return "opArgNameString"
-	case opArgSuperName:
-		return "opArgSuperName"
-	case opArgSimpleName:
-		return "opArgSimpleName"
-	case opArgDataRefObj:
-		return "opArgDataRefObj"
-	case opArgTarget:
-		return "opArgTarget"
-	case opArgFieldList:
-		return "opArgFieldList"
-	}
-	return ""
-}
-
-func makeArg0() opArgFlags                     { return 0 }
-func makeArg1(arg0 opArgFlag) opArgFlags       { return opArgFlags(arg0) }
-func makeArg2(arg0, arg1 opArgFlag) opArgFlags { return opArgFlags(arg1)<<8 | opArgFlags(arg0) }
-func makeArg3(arg0, arg1, arg2 opArgFlag) opArgFlags {
-	return opArgFlags(arg2)<<16 | opArgFlags(arg1)<<8 | opArgFlags(arg0)
-}
-func makeArg4(arg0, arg1, arg2, arg3 opArgFlag) opArgFlags {
-	return opArgFlags(arg3)<<24 | opArgFlags(arg2)<<16 | opArgFlags(arg1)<<8 | opArgFlags(arg0)
-}
-func makeArg5(arg0, arg1, arg2, arg3, arg4 opArgFlag) opArgFlags {
-	return opArgFlags(arg4)<<32 | opArgFlags(arg3)<<24 | opArgFlags(arg2)<<16 | opArgFlags(arg1)<<8 | opArgFlags(arg0)
-}
-func makeArg6(arg0, arg1, arg2, arg3, arg4, arg5 opArgFlag) opArgFlags {
-	return opArgFlags(arg5)<<40 | opArgFlags(arg4)<<32 | opArgFlags(arg3)<<24 | opArgFlags(arg2)<<16 | opArgFlags(arg1)<<8 | opArgFlags(arg0)
-}
-func makeArg7(arg0, arg1, arg2, arg3, arg4, arg5, arg6 opArgFlag) opArgFlags {
-	return opArgFlags(arg6)<<48 | opArgFlags(arg5)<<40 | opArgFlags(arg4)<<32 | opArgFlags(arg3)<<24 | opArgFlags(arg2)<<16 | opArgFlags(arg1)<<8 | opArgFlags(arg0)
-}
-
-// opcodeInfo contains all known information about an opcode,
-// its argument count and types as well as the type of object
-// represented by it.
-type opcodeInfo struct {
-	op      opcode
-	name    string
-	objType objType
-
-	flags    opFlag
-	argFlags opArgFlags
-}
